Add files via upload
The pickl will be located in the Google Drive directory, as that's the only place I can share it due to the size. Download this, it's all the data from the Greater Manchester area street dataset combined. There are 2 more datasets I can combine, but this is the main one, that we would have to start with. You do not have to run this notebook, but if you are curious how I merged all the files, go ahead and have a look. The last two cells save and check the pickl if it is correct.